# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreatePartnerTenantRequest:
    properties:
      name:
        maxLength: 150
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LogoutRequest:
    properties:
      refreshToken:
        type: string
    type: object
  dto.RefreshTokenRequest:
    properties:
      refreshToken:
        type: string
    required:
    - refreshToken
    type: object
  dto.RegisterRequest:
    properties:
      email:
        type: string
      fullName:
        maxLength: 150
        minLength: 2
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - fullName
    - password
    type: object
...
  response.Envelope:
    properties:
      data: {}
      error: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Backend API Service untuk platform SaaS POS & Self-Order UMKM kuliner.
  title: saas_gangsta Backend API
  version: "1.0"
paths:
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login untuk semua role (CUSTOMER, PARTNER, ADMIN)
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user dari sesi aktif
      parameters:
      - description: Logout payload
        in: body
        name: request
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - Auth
  /auth/me:
    get:
      description: Ambil profil user yang sedang login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Current user info
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh access token menggunakan refresh token yang valid
      parameters:
      - description: Refresh payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Refresh token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Membuat akun baru untuk login dengan role default customer
      parameters:
      - description: Register payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      summary: Register user account
      tags:
      - Auth
...
  /partner/tenants:
    get:
      description: Ambil daftar tenant milik partner login
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: List partner tenants
      tags:
      - Partner
    post:
      consumes:
      - application/json
      description: Partner membuat tenant baru miliknya sesuai limit paket subscription
      parameters:
      - description: Create partner tenant payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreatePartnerTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Envelope'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Envelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Envelope'
      security:
      - BearerAuth: []
      summary: Create partner tenant
      tags:
      - Partner
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"